handled the 'return with no value' warning in multiple driver modules

+++ STM32F407/Drivers/dma_STM32F407.h
@@ -14,7 +14,7 @@
  * @param dmaNum 
  * @return ** void 
  */
-void DMA_Init(int dmaNum);
+void DMA_Init(uint8_t dmaNum);
 
 //CLOCK
 #define CLOCK 0x40023800
